Simplystore

+++ www/assets/js/databrowser.js
@@ -133,10 +133,12 @@ browser = simply.app({
             browser.actions.item(params.id);
         },
 
-        '/register/': function(params) { 
+        '/register/': function(params) {
+            browser.actions.clearView()
             browser.view.view = 'register';
         },
         '/curriculum-:context/': function(params) {
+            browser.actions.clearView()
             browser.view.context = params.context;
             browser.view.contextLink = {
                 href: '/curriculum-'+params.context+'/',
@@ -156,6 +158,7 @@ browser = simply.app({
             updateDataSource('contextdata')
         },
         '/': function(params) {
+            browser.actions.clearView()
             browser.view.view = 'home';
         },
         '/([^#]+):*': function(params) {
@@ -461,6 +464,9 @@ browser = simply.app({
     },
 
     commands: {
+        toggleMaximize: (el, value) =>{
+            browser.view.sloSpreadsheet.selectorToggleMaximize()
+        },
         toggleSource: (el, value) => {
             browser.view.showSource = browser.view.showSource ? 0 : 1
             let url = new URL(document.location.href)
@@ -719,11 +725,13 @@ browser = simply.app({
         register : function(el, value) {
             var email = encodeURIComponent(el.querySelector("input").value);
             browser.actions.register(email);
+            browser.actions.clearView()
             browser.view.view = "registered";
         },
         'search-text': function(form, values) {
             const text = values.text;
             if (text) {
+                browser.actions.clearView()
                 browser.actions.search(text)
                 .then(data => {
                     browser.view.view = 'list';
@@ -746,10 +754,11 @@ browser = simply.app({
         },
         removeAllChanges: async function(el, value) {
             if (confirm('Weet u zeker dat u al uw lokale wijzigingen wil verwijderen?')) {
-                browser.actions.removeAllChanges()
-                browser.actions.switchView(browser.view.preferedView)
+                document.getElementById('previewChanges').showModal()
+                if (await browser.actions.removeAllChanges()) {
+                    browser.actions.switchView(browser.view.preferedView)
+                }
                 browser.commands.closeDialog(el, value)
-
             }
         },
         commitChanges: async function(form, values) {
@@ -928,6 +937,7 @@ browser = simply.app({
                 browser.view.loggedIn = false
                 return 'Ongeldig email en/of API-key'
             }
+            browser.actions.clearView()
             browser.view.user = email
             browser.view.loggedIn = true
             localStorage.setItem('username',email)
@@ -1129,6 +1139,7 @@ browser = simply.app({
             element.replaceChildren();
         },
         switchView: async function(view,root){
+            browser.actions.clearView()
             let currentView = this.app.view.view;
             let item = this.app.view.item
             let id = item?.id ?? item?.uuid
@@ -1140,6 +1151,7 @@ browser = simply.app({
             
             switch(view) {
                 case 'item':
+                    this.app.view.view = view
                     document.body.setAttribute('data-simply-keyboard','item')
                     this.app.actions.updatePaging()
                     // get focused item
@@ -1149,6 +1161,7 @@ browser = simply.app({
                 case 'spreadsheet':
                     document.body.setAttribute('data-simply-keyboard','spreadsheet')
                     this.app.actions.updatePaging()
+                    this.app.view.view = view
                     currentItem = id
                     currentId = 'https://opendata.slo.nl/curriculum/uuid/'+currentItem
                     if (!root) {
@@ -1260,6 +1273,7 @@ browser = simply.app({
                 case 'document':
                     document.body.setAttribute('data-simply-keyboard','document')
                     this.app.actions.updatePaging()
+                    this.app.view.view = view
                     currentItem = id
                     currentId = this.app.view.item['@id']
                     // get roots of current item
@@ -1506,6 +1520,7 @@ browser = simply.app({
             return true
         },
         list: function(type) {
+            browser.actions.clearView()
             browser.view['listTitle'] = window.titles[type];
             browser.view.list = [];
             return window.localAPI.list(type)
@@ -1546,6 +1561,7 @@ browser = simply.app({
         editText : function(){
         },
         spreadsheet: function(root, context, niveau) {
+            browser.actions.clearView()
             browser.view.currentTree = {
                 root,
                 context,
@@ -1591,6 +1607,7 @@ browser = simply.app({
             browser.view.sloSpreadsheet.render()
         },
         document: async function(root, context, niveau) {
+            browser.actions.clearView()
             browser.view.documentList = []
             return window.localAPI.document(root, context, niveau)
             .then(async function(json) {
@@ -1611,6 +1628,7 @@ browser = simply.app({
             })
         },
         doelniveauList: function(type) {
+            browser.actions.clearView()
             browser.view['listTitle'] = window.titles[type];
             browser.view.list = [];
             return window.localAPI.doelniveauList(type)
@@ -1627,6 +1645,7 @@ browser = simply.app({
             })
         },
         item: function(id) {
+            browser.actions.clearView()
             return window.localAPI.item(id)
             .then(function(json) {
                 browser.view.request = window.localAPI.reflect.item(id)
@@ -1645,6 +1664,7 @@ browser = simply.app({
             })
         },
         listOpNiveau: function(niveau, type) {
+            browser.actions.clearView()
             browser.view['listTitle'] = window.titles[type];
             browser.view.list = [];
             return window.localAPI.listOpNiveau(niveau, type)
@@ -1662,6 +1682,7 @@ browser = simply.app({
             })
         },
         itemOpNiveau: function(niveau, type, id) {
+            browser.actions.clearView()
             return window.localAPI.itemOpNiveau(niveau, type, id)
             .then(function(json) {
                 browser.view.source = JSON.stringify(json, null, 4)
@@ -1674,6 +1695,7 @@ browser = simply.app({
             })
         },
         notfound: function(remainder) {
+            browser.actions.clearView()
             browser.view.view = 'notfound';
             browser.actions.updatePaging();
         },
@@ -1687,6 +1709,9 @@ browser = simply.app({
                 browser.view.max = Math.ceil(pages);
             }
             MathJax.Hub.Queue(["Typeset",MathJax.Hub]);
+        },
+        clearView: function() {
+            document.body.classList.remove('ds-paging')
         },
         register : function(email) {
             window.slo.api.get("register/?email=" + email)
@@ -1785,28 +1810,44 @@ browser = simply.app({
         undeleteRow: function(rowEl) {
             row = browser.view.sloSpreadsheet.getRow(rowEl)
             if (row.deleted) {
+                let change
                 let parent = browser.view.sloSpreadsheet.findParentRow(row)
-                let parentNode = parent.node
-                let typeName = getType(row.node)
-                let prevValue = parentNode[typeName].slice()
-                let newValue = parentNode[typeName].slice()
-                row.node.undelete(newValue)
-                parentNode[typeName] = newValue //Makes sure current datamodel is up to date
-                let timestamp = new Date().toISOString()
-                let change = new changes.Change({
-                    id: parentNode.id ?? parentNode.uuid,
-                    meta: {
-                        context: window.slo.getContextByTypeName(getType(parentNode)),
-                        title: 'undelete child in '+parentNode.title,
-                        type: getType(parentNode),   
-                        timestamp: timestamp.substring(0, timestamp.indexOf('.'))
-                    },
-                    type: 'undelete',
-                    property: typeName,
-                    prevValue,
-                    newValue,
-                    dirty: true
-                })
+                if (parent && parent.node) {
+                    let parentNode = parent.node
+                    let typeName = getType(row.node)
+                    let prevValue = parentNode[typeName].slice()
+                    let newValue = parentNode[typeName].slice()
+                    row.node.undelete(newValue)
+                    parentNode[typeName] = newValue //Makes sure current datamodel is up to date
+                    let timestamp = new Date().toISOString()
+                    change = new changes.Change({
+                        id: parentNode.id ?? parentNode.uuid,
+                        meta: {
+                            context: window.slo.getContextByTypeName(getType(parentNode)),
+                            title: 'undelete child in '+parentNode.title,
+                            type: getType(parentNode),   
+                            timestamp: timestamp.substring(0, timestamp.indexOf('.'))
+                        },
+                        type: 'undelete',
+                        property: typeName,
+                        prevValue,
+                        newValue,
+                        dirty: true
+                    })
+                } else {
+                    let node = row.node
+                    let timestamp = new Date().toISOString()
+                    change = new changes.Change({
+                        id: node.id ?? node.uuid,
+                        meta: {
+                            context: window.slo.getContextByTypeName(getType(node)),
+                            title: 'undelete '+node.title,
+                            type: getType(node),
+                            timestamp: timestamp.substring(0, timestamp.indexOf('.'))
+                        },
+                        type: 'undeleteRoot'
+                    })
+                }
                 changes.changes.push(change)
                 changes.update()
                 //this.app.actions.switchView('spreadsheet')
@@ -1817,35 +1858,50 @@ browser = simply.app({
             if (!browser.view.user) return
             row = browser.view.sloSpreadsheet.getRow(row)
             let parent = browser.view.sloSpreadsheet.findParentRow(row)
-            let parentNode = parent.node
-            let typeName = getType(row.node)
+            let change 
+            if (parent && parent.node) {
+                let parentNode = parent.node
+                let typeName = getType(row.node)
 
-            let prevValue = parentNode[typeName].slice()
-            let newValue = prevValue.map(e => { 
-                if (e.id==row.node.id) {
-                    return Object.assign({}, e, {$mark:'deleted'}) // clone e, otherwise prevValue is changed as well
-                } else {
-                    return e
-                }
-            }) //filter(e => (e.id !== row.node.id))
-            let timestamp = new Date().toISOString()
-            let change = new changes.Change({
-                id: parentNode.id ?? parentNode.uuid,
-                meta: {
-                    context: window.slo.getContextByTypeName(getType(parentNode)),
-                    title: 'delete child of '+parentNode.title,
-                    type: getType(parentNode),
-                    timestamp: timestamp.substring(0, timestamp.indexOf('.'))
-                },
-                type: 'delete',
-                property: typeName,
-                prevValue,
-                newValue,
-                dirty: true,
-            })
+                let prevValue = parentNode[typeName].slice()
+                let newValue = prevValue.map(e => { 
+                    if (e.id==row.node.id) {
+                        return Object.assign({}, e, {$mark:'deleted'}) // clone e, otherwise prevValue is changed as well
+                    } else {
+                        return e
+                    }
+                }) //filter(e => (e.id !== row.node.id))
+                let timestamp = new Date().toISOString()
+                change = new changes.Change({
+                    id: parentNode.id ?? parentNode.uuid,
+                    meta: {
+                        context: window.slo.getContextByTypeName(getType(parentNode)),
+                        title: 'delete child of '+parentNode.title,
+                        type: getType(parentNode),
+                        timestamp: timestamp.substring(0, timestamp.indexOf('.'))
+                    },
+                    type: 'delete',
+                    property: typeName,
+                    prevValue,
+                    newValue,
+                    dirty: true,
+                })
+            } else { // rootnode
+                let node = row.node
+                let timestamp = new Date().toISOString()
+                change = new changes.Change({
+                    id: node.id ?? node.uuid,
+                    meta: {
+                        context: window.slo.getContextByTypeName(getType(node)),
+                        title: 'delete '+node.title,
+                        type: getType(node),
+                        timestamp: timestamp.substring(0, timestamp.indexOf('.'))
+                    },
+                    type: 'deleteRoot'
+                })
+            }
             changes.changes.push(change)
             changes.update()
-//            this.app.actions.switchView('spreadsheet')
             browser.actions.spreadsheetUpdate()
         },
         showTypeSelector: async function(el) {
@@ -1909,6 +1965,29 @@ browser = simply.app({
         },
         removeAllChanges: async function() {
             changes.clear()
+            return browser.actions.checkView()
+        },
+        checkView: async function() {
+            // check if the current view can be updated,
+            // or if browser should show the homepage instead
+            // e.g. current uuid was a local change, which has been removed
+            switch(browser.view.view) {
+                case 'item':
+                case 'spreadsheet':
+                case 'document':
+                    let uuid = browser.view.item.id
+                    try {
+                        let item = await localAPI.item(uuid)
+                    } catch(e) {
+                        browser.actions.clearView()
+                        simply.route.goto('/')
+                        return false
+                    }
+                break
+            }
+            // now update current view
+            window.location.reload(true)
+            return true
         },
         commitChanges: async function(message) {
             const linkArray = (list) => {
